# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clarity - Document Understanding API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/profile:
    patch:
      operationId: updateProfile
      tags: [settings]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [free, pro]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - plan
        - createdAt
...
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
...